requirements.txt
# LLM Engine Dependencies

# Core LLM and Model Loading
llama-cpp-python>=0.2.0
transformers>=4.30.0
torch>=2.0.0

# Embeddings and Reranking
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.4.0

# Terminal UI
rich>=13.0.0

# Configuration
pyyaml>=6.0

# System Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# Data Processing
numpy>=1.24.0

# File Monitoring (for autonomous mode)
watchdog>=3.0.0

# Data Validation
pydantic>=2.0.0

# Testing (development)
pytest>=7.4.0
